update: storybook and deps
So there was an issue with storybook. The error indicates that Vite’s PostCSS pretransform is mistakenly processing non‑CSS content (the `<script>` tag) from the Svelte file instead of just the CSS, causing the "Unknown word" error. I upgraded most deps but also left some in older versions because I am not ready to make that large of a migration. Maybe in a couple months I can move to the latest version of vite and svelte.